Basic server setup instructions.
There isn't really much to this project, so hopefully it is easy enough
to set up wherever. The instructions are meant for a Raspberry Pi
running Raspbian, but really any Linux server you have root access to
would probably work. To get it working on a Windows server... you'll
have to figure out all the equivalent setup. Sorry.